Fix TSC build and linting
Update TypeScript & related configs and code so TSC can actually be run
against the codebase without errors.

The main tradeoff that had to be made was to stop using top level await
statements in scripts, as using `"type": "module"` in package.json
caused the r2modmanPlus imports to not function properly.

In the future once the r2modmanPlus submodule is removed, the
aforementioned change can be reverted.