Update deployment script to upload sequentially
There's a serverside race condition bug with the deployment endpoint,
where if two identical files are uploaded in parallel, one of the
uploads will fail (even if release channels are different).

As a temporary workaround convert the deployment process to uploading
files sequentially. Parallel uploads can be re-enabled once the
serverside bug has been addressed, but as it requires distributed
locking, it's likely going to take longer than a workaround on the
client now.