Add support for PackageLoader.None
PackageLoader.None was added to provide generic support for games that
don't use any package loaders. Null value is no longer accepted, a
valid package loader enum value needs to be used instead. This creates
a bit of chicken/egg problem as the current r2mm develop branch still
validly support the null value, but at the same time we don't want to
support it at this end when extacting values from r2mm. This is
bypassed with a simple casting to avoid spending too much effort
working on this temporary code.