Game addition script: prompt for mod manager support
Add prompts for filling in the data to add support for 90%+ of the
cases. For the uncommon cases, the resulting yml file can be edited
manually.

There was a conflict between the TypeScript types and the fields
checked by the validator script - the types were updated to resolve the
issue.