Game addition script: fill in default install rules when using BepInEx
Older MelonLoader implementation's install rules are not applied as
hopefully it's not used anymore now that recursive MelonLoader provides
a better developer experience.

For GodotML and Northstar it was difficult to figure whether the
existing rules are general enough that they can be considered defaults,
as there was only few instances of use.