Add support for UMM and Broforce
Unity Mod Manager is a in-game mod manager and loader. The version used
with Thunderstore is a fork where manager functionality is limited to
enabling/disabling mods in a in-game window. Downloads and installation
of mods is handled by TSMM or compatible mod manager. UMM acts mostly
as the mod loader.

UMM is used since the game's community had already established around
it and switching to another mod loader wasn't seen as feasible.